+++ src/lib.rs
@@ -1,261 +1,36 @@
 use once_cell::sync::OnceCell;
 use rustc_hash::FxHashMap;
 use spin::Mutex;
-use std::{collections::hash_map::Entry, ffi::c_void};
+use std::{collections::hash_map::Entry, ffi::c_void, num::NonZeroUsize};
 
 #[cfg(all(target_arch = "x86", target_os = "windows", target_env = "gnu"))]
 #[allow(unused)]
 #[no_mangle]
 unsafe extern "C" fn _Unwind_Resume() {}
 
-static LEDGER: OnceCell<Mutex<Ledger>> = OnceCell::new();
-
 /// The current API version.
 #[no_mangle]
-pub static JLRS_LEDGER_API_VERSION: usize = 2;
+pub static JLRS_LEDGER_API_VERSION: usize = 3;
 
-/// Indicates whether an operation succeeded or failed.
+const EXCLUSIVE_BORROW_MASK: usize = 1;
+const EXCLUSIVE_BORROW: NonZeroUsize =
+    unsafe { NonZeroUsize::new_unchecked(EXCLUSIVE_BORROW_MASK) };
+const SHARED_BORROW_MASK: usize = 2;
+const SHARED_BORROW: NonZeroUsize = unsafe { NonZeroUsize::new_unchecked(SHARED_BORROW_MASK) };
+
+static LEDGER: OnceCell<Mutex<FxHashMap<usize, NonZeroUsize>>> = OnceCell::new();
+
+/// Indicates whether an operation succeeded, failed, or incorrect.
 #[derive(PartialEq, Eq, Debug)]
-#[repr(u8)]
+#[repr(i32)]
 pub enum LedgerResult {
-    /// Operation succeeded, result is `false`.
-    OkFalse = 0u8,
+    /// Operation succeeded and the result is `false`, or the operation failed.
+    OkFalse = 0,
     /// Operation succeeded, result is `true`.
-    OkTrue = 1u8,
-    /// Operation failed.
-    Err = 2u8,
+    OkTrue = 1,
+    /// Operation is incorrect given the state of the ledger.
+    Err = -1,
 }
-
-/// Tracker for Julia data that is borrowed in Rust code.
-pub struct Ledger {
-    state: FxHashMap<usize, usize>,
-}
-
-impl Ledger {
-    #[inline(always)]
-    fn new() -> Self {
-        Ledger {
-            state: Default::default(),
-        }
-    }
-
-    /// Initialize the ledger. Does nothing if the ledger has already been initialized.
-    #[inline(always)]
-    pub fn init() {
-        LEDGER.get_or_init(|| Mutex::new(Ledger::new()));
-    }
-
-    /// Check if the given pointer is tracked as a shared borrow.
-    ///
-    /// Returns `LedgerResult::OkTrue` if it is borrowed sharedly, or `LedgerResult::OkFalse` if
-    /// it isn't.
-    ///
-    /// Safety:
-    ///
-    /// `Ledger::init` must have been called before calling this function.
-    #[inline(always)]
-    pub unsafe fn is_borrowed_shared(ptr: *const c_void) -> LedgerResult {
-        let ledger = unsafe { LEDGER.get_unchecked().lock() };
-        let borrowed_shared = if let Some(state) = ledger.state.get(&(ptr as usize)) {
-            state != &1
-        } else {
-            false
-        };
-
-        if borrowed_shared {
-            return LedgerResult::OkTrue;
-        } else {
-            return LedgerResult::OkFalse;
-        }
-    }
-
-    /// Check if the given pointer is tracked as an exclusive borrow.
-    ///
-    /// Returns `LedgerResult::OkTrue` if it is borrowed exclusively, or `LedgerResult::OkFalse`
-    /// if it isn't.
-    ///
-    /// Safety:
-    ///
-    /// `Ledger::init` must have been called before calling this function.
-    #[inline(always)]
-    pub unsafe fn is_borrowed_exclusive(ptr: *const c_void) -> LedgerResult {
-        let ledger = unsafe { LEDGER.get_unchecked().lock() };
-        let borrowed_exclusive = if let Some(state) = ledger.state.get(&(ptr as usize)) {
-            state == &1
-        } else {
-            false
-        };
-
-        if borrowed_exclusive {
-            return LedgerResult::OkTrue;
-        } else {
-            return LedgerResult::OkFalse;
-        }
-    }
-
-    /// Check if the given pointer is tracked in some way.
-    ///
-    /// Returns `LedgerResult::OkTrue` if it is borrowed, or `LedgerResult::OkFalse`
-    /// if it isn't.
-    ///
-    /// Safety:
-    ///
-    /// `Ledger::init` must have been called before calling this function.
-    #[inline(always)]
-    pub unsafe fn is_borrowed(ptr: *const c_void) -> LedgerResult {
-        let ledger = unsafe { LEDGER.get_unchecked().lock() };
-
-        if ledger.state.contains_key(&(ptr as usize)) {
-            return LedgerResult::OkTrue;
-        } else {
-            return LedgerResult::OkFalse;
-        }
-    }
-
-    /// Try to track the pointer as a shared borrow.
-    ///
-    /// Returns `LedgerResult::Ok` on success, or `LedgerResult:Err` if an exclusive borrow
-    /// already exists.
-    ///
-    /// Safety:
-    ///
-    /// `Ledger::init` must have been called before calling this function.
-    #[inline(always)]
-    pub unsafe fn try_borrow_shared(ptr: *const c_void) -> LedgerResult {
-        LEDGER.get_unchecked().lock().try_add_borrow_shared(ptr)
-    }
-
-    /// Try to track the pointer as an exclusive borrow.
-    ///
-    /// Returns `LedgerResult::Ok` on success, or `LedgerResult:Err` if an exclusive borrow
-    /// already exists.
-    ///
-    /// Safety:
-    ///
-    /// `Ledger::init` must have been called before calling this function.
-    #[inline(always)]
-    pub unsafe fn try_borrow_exclusive(ptr: *const c_void) -> LedgerResult {
-        LEDGER.get_unchecked().lock().try_add_borrow_exclusive(ptr)
-    }
-
-    /// Track the pointer as a shared borrow without checking if it is already borrowed.
-    ///
-    /// Returns `LedgerResult::Ok`.
-    ///
-    /// Safety:
-    ///
-    /// `Ledger::init` must have been called before calling this function. The pointer must not
-    /// already be tracked as an exclusive borrow.
-    #[inline(always)]
-    pub unsafe fn borrow_shared_unchecked(ptr: *const c_void) -> LedgerResult {
-        LEDGER
-            .get_unchecked()
-            .lock()
-            .state
-            .entry(ptr as usize)
-            .and_modify(|i| *i += 2)
-            .or_insert(2);
-        LedgerResult::OkTrue
-    }
-
-    /// Stop tracking the pointer as a shared borrow.
-    ///
-    /// Returns `LedgerResult::OkTrue` if it's no longer tracked, `LedgerResult::OkFalse` if other
-    /// shared borrows still exist, and `LedgerResult::Err` if the pointer wasn't tracked.
-    ///
-    /// Safety:
-    ///
-    /// `Ledger::init` must have been called before calling this function. The pointer must
-    /// have been tracked as a shared borrow. A `LedgerResult::Err` being returned by this
-    /// function implies that the ledger has been corrupted.
-    #[inline(always)]
-    pub unsafe fn unborrow_shared(ptr: *const c_void) -> LedgerResult {
-        let mut ledger = LEDGER.get_unchecked().lock();
-        let entry = ledger.state.entry(ptr as usize);
-
-        match entry {
-            Entry::Occupied(mut o) => {
-                if o.get() == &2 {
-                    o.remove_entry();
-                    LedgerResult::OkTrue
-                } else {
-                    *o.get_mut() -= 2;
-                    LedgerResult::OkFalse
-                }
-            }
-            Entry::Vacant(_) => LedgerResult::Err,
-        }
-    }
-
-    /// Stop tracking the pointer as an exclusive borrow.
-    ///
-    /// Returns `LedgerResult::OkTrue` if it's no longer tracked, and `LedgerResult::Err` if the
-    /// pointer wasn't tracked or was tracked as a shared borrow.
-    ///
-    /// Safety:
-    ///
-    /// `Ledger::init` must have been called before calling this function. The pointer must
-    /// have been tracked as an exclusive borrow. A `LedgerResult::Err` being returned by this
-    /// function implies that the ledger has been corrupted.
-    #[inline(always)]
-    pub unsafe fn unborrow_exclusive(ptr: *const c_void) -> LedgerResult {
-        let mut ledger = LEDGER.get_unchecked().lock();
-        if let Some(v) = ledger.state.remove(&(ptr as usize)) {
-            if v == 1 {
-                LedgerResult::OkTrue
-            } else {
-                LedgerResult::Err
-            }
-        } else {
-            LedgerResult::Err
-        }
-    }
-
-    /// Clear the entire ledger.
-    ///
-    /// Don't call this function. It only exists to clear the ledger between benchmarks and tests.
-    #[doc(hidden)]
-    #[inline(always)]
-    pub unsafe fn clear() {
-        let mut ledger = LEDGER.get_unchecked().lock();
-        ledger.state.clear();
-    }
-
-    #[inline(always)]
-    // Try to add an immutable borrow to the ledger
-    unsafe fn try_add_borrow_shared(&mut self, ptr: *const c_void) -> LedgerResult {
-        // Check if the data is already borrowed
-        let entry = self.state.entry(ptr as usize);
-        match &entry {
-            Entry::Occupied(o) => {
-                if o.get() == &1 {
-                    return LedgerResult::Err;
-                }
-            }
-            _ => (),
-        }
-
-        // Push the pointer to indicate it's borrowed
-        entry.and_modify(|i| *i += 2).or_insert(2);
-        LedgerResult::OkTrue
-    }
-
-    #[inline(always)]
-    // Try to add a mutable borrow to the ledger
-    unsafe fn try_add_borrow_exclusive(&mut self, ptr: *const c_void) -> LedgerResult {
-        // Check if the borrow overlaps with any active mutable borrow
-        let entry = self.state.entry(ptr as usize);
-        match &entry {
-            Entry::Occupied(_) => LedgerResult::Err,
-            _ => {
-                entry.or_insert(1);
-                LedgerResult::OkTrue
-            }
-        }
-    }
-}
-
-/* C-API */
 
 /// Return the API version.
 #[no_mangle]
@@ -265,64 +40,186 @@ pub unsafe extern "C" fn jlrs_ledger_api_version() -> usize {
 
 /// Initialize the ledger.
 ///
-/// This function must be called before calling any other functions from the C-API of this
-/// library, except [`jlrs_ledger_api_version`].
+/// This function must be called before calling any other functions from the C API of this
+/// crate, except [`jlrs_ledger_api_version`].
 #[no_mangle]
-pub unsafe extern "C" fn jlrs_ledger_init() {
-    Ledger::init();
+pub extern "C" fn jlrs_ledger_init() {
+    LEDGER.get_or_init(|| Mutex::new(Default::default()));
 }
 
-/// Call [`Ledger::is_borrowed_shared`].
+/// Check if the given pointer is tracked as a shared borrow.
+///
+/// Returns `LedgerResult::OkTrue` if it is borrowed sharedly, or `LedgerResult::OkFalse` if
+/// it isn't.
+///
+/// Safety:
+///
+/// `jlrs_ledger_init` must have been called before calling this function.
 #[no_mangle]
 pub unsafe extern "C" fn jlrs_ledger_is_borrowed_shared(ptr: *const c_void) -> LedgerResult {
-    Ledger::is_borrowed_shared(ptr)
+    match LEDGER.get_unchecked().lock().get(&(ptr as usize)).copied() {
+        None => LedgerResult::OkFalse,
+        Some(s) if s != EXCLUSIVE_BORROW => LedgerResult::OkTrue,
+        Some(_) => LedgerResult::OkFalse,
+    }
 }
 
-/// Call [`Ledger::is_borrowed_exclusive`].
+/// Check if the given pointer is tracked as an exclusive borrow.
+///
+/// Returns `LedgerResult::OkTrue` if it is borrowed exclusively, or `LedgerResult::OkFalse`
+/// if it isn't.
+///
+/// Safety:
+///
+/// `jlrs_ledger_init` must have been called before calling this function.
 #[no_mangle]
 pub unsafe extern "C" fn jlrs_ledger_is_borrowed_exclusive(ptr: *const c_void) -> LedgerResult {
-    Ledger::is_borrowed_exclusive(ptr)
+    match LEDGER.get_unchecked().lock().get(&(ptr as usize)).copied() {
+        None => LedgerResult::OkFalse,
+        Some(s) if s == EXCLUSIVE_BORROW => LedgerResult::OkTrue,
+        Some(_) => LedgerResult::OkFalse,
+    }
 }
 
-/// Call [`Ledger::is_borrowed`].
+/// Check if the given pointer is tracked in some way.
+///
+/// Returns `LedgerResult::OkTrue` if it is borrowed, or `LedgerResult::OkFalse`
+/// if it isn't.
+///
+/// Safety:
+///
+/// `jlrs_ledger_init` must have been called before calling this function.
 #[no_mangle]
 pub unsafe extern "C" fn jlrs_ledger_is_borrowed(ptr: *const c_void) -> LedgerResult {
-    Ledger::is_borrowed(ptr)
+    if LEDGER.get_unchecked().lock().contains_key(&(ptr as usize)) {
+        LedgerResult::OkTrue
+    } else {
+        LedgerResult::OkFalse
+    }
 }
 
-/// Call [`Ledger::borrow_shared_unchecked`].
+/// The number of active shared borrows of `addr`.
+///
+/// Safety:
+///
+/// `jlrs_ledger_init` must have been called before calling this function.
 #[no_mangle]
-pub unsafe extern "C" fn jlrs_ledger_borrow_shared_unchecked(ptr: *const c_void) -> LedgerResult {
-    Ledger::borrow_shared_unchecked(ptr)
+pub unsafe extern "C" fn jlrs_ledger_n_shared_borrows(ptr: *const c_void) -> usize {
+    match LEDGER.get_unchecked().lock().get(&(ptr as usize)).copied() {
+        Some(s) => s.get() >> 1,
+        None => 0,
+    }
 }
-
-/// Call [`Ledger::unborrow_shared`].
+/// Stop tracking the pointer as a shared borrow.
+///
+/// Returns `LedgerResult::OkTrue` if it's no longer tracked, `LedgerResult::OkFalse` if other
+/// shared borrows still exist, and `LedgerResult::Err` if the pointer wasn't tracked or is
+/// tracked as an exclusive borrow.
+///
+/// Safety:
+///
+/// `jlrs_ledger_init` must have been called before calling this function.
 #[no_mangle]
 pub unsafe extern "C" fn jlrs_ledger_unborrow_shared(ptr: *const c_void) -> LedgerResult {
-    Ledger::unborrow_shared(ptr)
+    match LEDGER.get_unchecked().lock().entry(ptr as usize) {
+        Entry::Occupied(mut o) => {
+            let current = *o.get();
+            if current == SHARED_BORROW {
+                o.remove_entry();
+                LedgerResult::OkTrue
+            } else if current != EXCLUSIVE_BORROW {
+                *o.get_mut() = NonZeroUsize::new_unchecked(current.get() - SHARED_BORROW_MASK);
+                LedgerResult::OkFalse
+            } else {
+                LedgerResult::Err
+            }
+        }
+        Entry::Vacant(_) => LedgerResult::Err,
+    }
 }
 
-/// Call [`Ledger::unborrow_exclusive`].
+/// Stop tracking the pointer as an exclusive borrow.
+///
+/// Returns `LedgerResult::OkTrue` if it's no longer tracked, and `LedgerResult::Err` if the
+/// pointer wasn't tracked or was tracked as a shared borrow.
+///
+/// Safety:
+///
+/// `jlrs_ledger_init` must have been called before calling this function.
 #[no_mangle]
 pub unsafe extern "C" fn jlrs_ledger_unborrow_exclusive(ptr: *const c_void) -> LedgerResult {
-    Ledger::unborrow_exclusive(ptr)
+    let mut ledger = LEDGER.get_unchecked().lock();
+    match ledger.remove(&(ptr as usize)) {
+        Some(s) if s == EXCLUSIVE_BORROW => LedgerResult::OkTrue,
+        Some(s) => {
+            ledger.insert(ptr as _, s);
+            LedgerResult::Err
+        }
+        _ => LedgerResult::Err,
+    }
 }
 
-/// Call [`Ledger::try_borrow_shared`].
+/// Try to track the pointer as a shared borrow.
+///
+/// Returns `LedgerResult::OkTrue` on success, or `LedgerResult:OkFalse` if it is already
+/// tracked as an exclusive borrow.
+///
+/// Safety:
+///
+/// `jlrs_ledger_init` must have been called before calling this function.
 #[no_mangle]
 pub unsafe extern "C" fn jlrs_ledger_try_borrow_shared(ptr: *const c_void) -> LedgerResult {
-    Ledger::try_borrow_shared(ptr)
+    match LEDGER.get_unchecked().lock().entry(ptr as usize) {
+        Entry::Vacant(v) => {
+            v.insert(SHARED_BORROW);
+            LedgerResult::OkTrue
+        }
+        Entry::Occupied(ref mut o) => {
+            let current = *o.get();
+            if current != EXCLUSIVE_BORROW {
+                *o.get_mut() = NonZeroUsize::new_unchecked(current.get() + SHARED_BORROW_MASK);
+                LedgerResult::OkTrue
+            } else {
+                LedgerResult::OkFalse
+            }
+        }
+    }
 }
-
-/// Call [`Ledger::try_borrow_exclusive`].
+/// Try to track the pointer as an exclusive borrow.
+///
+/// Returns `LedgerResult::Ok` on success, or `LedgerResult:OkFalse` if it is already
+/// tracked.
+///
+/// Safety:
+///
+/// `jlrs_ledger_init` must have been called before calling this function.
 #[no_mangle]
 pub unsafe extern "C" fn jlrs_ledger_try_borrow_exclusive(ptr: *const c_void) -> LedgerResult {
-    Ledger::try_borrow_exclusive(ptr)
+    match LEDGER.get_unchecked().lock().entry(ptr as usize) {
+        Entry::Vacant(v) => {
+            v.insert(EXCLUSIVE_BORROW);
+            LedgerResult::OkTrue
+        }
+        Entry::Occupied(_) => LedgerResult::OkFalse,
+    }
+}
+
+/// Clear the ledger.
+///
+/// Don't call this function, it exists purely for testing and benchmarking purposes. It's not
+/// part of the C API of this crate.
+///
+/// Safety:
+///
+/// `jlrs_ledger_init` must have been called before calling this function.
+#[doc(hidden)]
+pub unsafe fn clear_ledger() {
+    LEDGER.get_unchecked().lock().clear();
 }
 
 #[cfg(test)]
 mod tests {
-    use std::ptr::NonNull;
+    use std::ptr::dangling;
 
     use super::*;
 
@@ -336,11 +233,12 @@ mod tests {
     #[test]
     fn borrow_shared() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_is_borrowed_shared(ptr), LedgerResult::OkFalse);
+            assert_eq!(jlrs_ledger_n_shared_borrows(ptr), 0);
             assert_eq!(
                 jlrs_ledger_is_borrowed_exclusive(ptr),
                 LedgerResult::OkFalse
@@ -348,6 +246,7 @@ mod tests {
             assert_eq!(jlrs_ledger_is_borrowed(ptr), LedgerResult::OkFalse);
             assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_is_borrowed_shared(ptr), LedgerResult::OkTrue);
+            assert_eq!(jlrs_ledger_n_shared_borrows(ptr), 1);
             assert_eq!(
                 jlrs_ledger_is_borrowed_exclusive(ptr),
                 LedgerResult::OkFalse
@@ -359,10 +258,10 @@ mod tests {
     #[test]
     fn borrow_exclusive() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_is_borrowed_shared(ptr), LedgerResult::OkFalse);
             assert_eq!(
                 jlrs_ledger_is_borrowed_exclusive(ptr),
@@ -371,6 +270,7 @@ mod tests {
             assert_eq!(jlrs_ledger_is_borrowed(ptr), LedgerResult::OkFalse);
             assert_eq!(jlrs_ledger_try_borrow_exclusive(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_is_borrowed_shared(ptr), LedgerResult::OkFalse);
+            assert_eq!(jlrs_ledger_n_shared_borrows(ptr), 0);
             assert_eq!(jlrs_ledger_is_borrowed_exclusive(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_is_borrowed(ptr), LedgerResult::OkTrue);
         }
@@ -379,49 +279,53 @@ mod tests {
     #[test]
     fn borrow_exclusive_then_shared() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_try_borrow_exclusive(ptr), LedgerResult::OkTrue);
-            assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::Err);
+            assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::OkFalse);
+            assert_eq!(jlrs_ledger_n_shared_borrows(ptr), 0);
         }
     }
 
     #[test]
     fn borrow_shared_then_exclusive() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::OkTrue);
-            assert_eq!(jlrs_ledger_try_borrow_exclusive(ptr), LedgerResult::Err);
+            assert_eq!(jlrs_ledger_try_borrow_exclusive(ptr), LedgerResult::OkFalse);
+            assert_eq!(jlrs_ledger_n_shared_borrows(ptr), 1);
         }
     }
 
     #[test]
     fn borrow_shared_twice() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::OkTrue);
+            assert_eq!(jlrs_ledger_n_shared_borrows(ptr), 2);
         }
     }
 
     #[test]
     fn unborrow_shared() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_unborrow_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_is_borrowed_shared(ptr), LedgerResult::OkFalse);
+            assert_eq!(jlrs_ledger_n_shared_borrows(ptr), 0);
             assert_eq!(
                 jlrs_ledger_is_borrowed_exclusive(ptr),
                 LedgerResult::OkFalse
@@ -433,10 +337,10 @@ mod tests {
     #[test]
     fn unborrow_exclusive() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_try_borrow_exclusive(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_unborrow_exclusive(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_is_borrowed_shared(ptr), LedgerResult::OkFalse);
@@ -451,13 +355,14 @@ mod tests {
     #[test]
     fn unborrow_shared_twice() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_unborrow_shared(ptr), LedgerResult::OkFalse);
+            assert_eq!(jlrs_ledger_n_shared_borrows(ptr), 1);
             assert_eq!(jlrs_ledger_is_borrowed_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(
                 jlrs_ledger_is_borrowed_exclusive(ptr),
@@ -466,6 +371,7 @@ mod tests {
             assert_eq!(jlrs_ledger_is_borrowed(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_unborrow_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_is_borrowed_shared(ptr), LedgerResult::OkFalse);
+            assert_eq!(jlrs_ledger_n_shared_borrows(ptr), 0);
             assert_eq!(
                 jlrs_ledger_is_borrowed_exclusive(ptr),
                 LedgerResult::OkFalse
@@ -477,10 +383,10 @@ mod tests {
     #[test]
     fn unborrow_unborrowed_shared() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_try_borrow_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_unborrow_shared(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_unborrow_shared(ptr), LedgerResult::Err);
@@ -490,10 +396,10 @@ mod tests {
     #[test]
     fn unborrow_unborrowed_exclusive() {
         unsafe {
-            Ledger::init();
-            Ledger::clear();
+            jlrs_ledger_init();
+            clear_ledger();
 
-            let ptr = NonNull::<c_void>::dangling().as_ptr();
+            let ptr = dangling();
             assert_eq!(jlrs_ledger_try_borrow_exclusive(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_unborrow_exclusive(ptr), LedgerResult::OkTrue);
             assert_eq!(jlrs_ledger_unborrow_exclusive(ptr), LedgerResult::Err);